feature/renderVudeoBlock and menu
