+++ src/pages/main/main.jsx
@@ -1,4 +1,3 @@
-
 import { useDispatch } from "react-redux";
 import * as S from "./mainStyled";
 import { Link } from "react-router-dom";
@@ -6,6 +5,7 @@ import { useEffect } from "react";
 import { setLogo } from "../../store/slices/logoSlices";
 import { Header } from "../../Components/header/header";
 import { useUser } from "../../firebase/getUser";
+import { useDataWorkout } from "../../firebase/fireWorkouts";
 
 const courses = [
   { id: "1", img: "/img/profCard1.png" },
@@ -16,18 +16,22 @@ const courses = [
 ];
 
 export const Main = () => {
-const dispatch = useDispatch();
-useEffect(() => {
-  dispatch(setLogo({
-    logo: "white",
-  }))
-}, []);
-useUser()
+  const dispatch = useDispatch();
+  useEffect(() => {
+    dispatch(
+      setLogo({
+        logo: "white",
+      })
+    );
+  }, []);
+  useUser();
 
-      return (
+  useDataWorkout();
+
+  return (
     <S.Container>
       <S.Content>
-        <Header/>
+        <Header />
         <S.HeaderContent>
           <S.SubTitle>
             <S.TitleText name="top">
@@ -52,6 +56,6 @@ useUser()
           </a>
         </S.Button>
       </S.Content>
-      </S.Container>
-      )
-    }
+    </S.Container>
+  );
+};
